just replace existing `rs` with tonic
