+++ Cargo.toml
@@ -1,5 +1,4 @@
 [workspace]
 members = [
-    "rs",
-    "rs-tonic",
+    "rs"
 ]
