--- rs-tonic/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "linkerd2-proxy-api-tonic"
-version = "0.1.12"
-authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
-publish = false
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-# Generate code that is compatible with Tonic's `transport` module.
-transport = ["tonic-build/transport"]
-# Enable generation of arbitrary protos with quickcheck.
-arbitrary = ["quickcheck", "rand"]
-
-[dependencies]
-tonic = { version = "0.2", default-features = false, features = ["prost", "codegen"] }
-prost = "0.6"
-prost-types = "0.6"
-h2 = "0.2"
-http = "0.2"
-
-# For `arbitrary`:
-quickcheck = { version = "0.9", default-features = false, optional = true }
-rand = { version = "0.7", optional = true }
-
-[build-dependencies]
-tonic-build = { version = "0.2", features = ["prost"], default-features = false }
