+++ rs/Cargo.toml
@@ -5,23 +5,24 @@ authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
 publish = false
 edition = "2018"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
 [features]
-default = []
+# Generate code that is compatible with Tonic's `transport` module.
+transport = ["tonic-build/transport"]
+# Enable generation of arbitrary protos with quickcheck.
 arbitrary = ["quickcheck", "rand"]
 
 [dependencies]
-bytes = "0.4"
-futures = "0.1"
-h2 = "0.1"
-http = "0.1"
-prost = "0.5"
-prost-types = "0.5"
-tower-grpc = "0.1"
+tonic = { version = "0.2", default-features = false, features = ["prost", "codegen"] }
+prost = "0.6"
+prost-types = "0.6"
+h2 = "0.2"
+http = "0.2"
 
 # For `arbitrary`:
 quickcheck = { version = "0.9", default-features = false, optional = true }
 rand = { version = "0.7", optional = true }
 
 [build-dependencies]
-tower-grpc-build = "0.1"
-
+tonic-build = { version = "0.2", features = ["prost"], default-features = false }
