+++ rs/build.rs
@@ -1,27 +1,28 @@
-extern crate tower_grpc_build;
-
-fn main() {
-    build_control();
+use std::error::Error;
+fn main() -> Result<(), Box<dyn Error>> {
+    build_control()?;
+    Ok(())
 }
 
-fn build_control() {
+fn build_control() -> Result<(), Box<dyn Error>> {
     let iface_files = &[
         "../proto/destination.proto",
         "../proto/http_types.proto",
         "../proto/identity.proto",
         "../proto/net.proto",
-        "../proto/tap.proto"
+        "../proto/tap.proto",
     ];
     let dirs = &["../proto"];
 
-    tower_grpc_build::Config::new()
-        .enable_client(true)
-        .enable_server(true)
-        .build(iface_files, dirs)
-        .unwrap();
+    tonic_build::configure()
+        .build_client(true)
+        .build_server(true)
+        .compile(iface_files, dirs)?;
 
     // recompile protobufs only if any of the proto files changes.
     for file in iface_files {
         println!("cargo:rerun-if-changed={}", file);
     }
+
+    Ok(())
 }
